refactor(utils): consolidate utility functions and remove unused code
- Move `ifisdir_rmdir` and `return_cutoff_values` functions to `blech_utils.py` for centralized utility management.
- Delete `blech_general_utils.py` as it is no longer needed after the refactor.
- Remove commented-out `generate_github_summary` code from `blech_utils.py` as it is unused and unnecessary.
